Title/GameOver Logic and Richer Enemy A.I.
I have completed some basic logic for when the player is on the title and game over screens. That way when were ready to implement the design, we already have the logic. I have also created enemy logic to make the enemy have different ways of performing actions using a random number generator.